runner/requirements.txt
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
snap-stanford==6.0.0
